perf(rdna3): fp16 cast for INT4 packed QK dot → WMMA activation
Cast Q_s0/K_s0 to fp16 before tl.dot to ensure Triton lowers to
v_wmma_f32_16x16x16_f16 instead of scalar FMA. Without the cast,
Triton's fp32 dot bypasses WMMA entirely on RDNA3.

TP2 Qwen3.6-27B at 32K context:
  Decode: 37.3 → 40.0 tok/s (+7%)
  Prefill: 1061 → 1623 tok/s (+53%)
  Short decode: 54.7 → 51.5 tok/s (-6%, precision trade-off)

+++ vllm/v1/attention/ops/triton_per_token_head_attention.py
@@ -819,7 +819,10 @@ def _pth_attn_stage1_packed_gqa(
             k_scales = ks_raw
 
         if PACKING_FACTOR == 2:
-            qk = tl.dot(Q_s0, K_s0) + tl.dot(Q_s1, K_s1)
+            # Cast to fp16 before dot to ensure WMMA (v_wmma_f32_16x16x16_f16)
+            # instead of scalar FMA. +7% decode, +53% prefill at 32K context.
+            qk = (tl.dot(Q_s0.to(tl.float16), K_s0.to(tl.float16))
+                  + tl.dot(Q_s1.to(tl.float16), K_s1.to(tl.float16)))
             qk = (qk - Q_sum[:, None] * k_zp[None, :]) * (
                 ATTN_SCALE * k_scales[None, :]
             )
